+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/ConfigurationImpl.kt
@@ -39,10 +39,10 @@ import io.realm.kotlin.internal.platform.appFilesDirectory
 import io.realm.kotlin.internal.platform.freeze
 import io.realm.kotlin.internal.platform.prepareRealmFilePath
 import io.realm.kotlin.internal.platform.realmObjectCompanionOrThrow
+import io.realm.kotlin.internal.util.CoroutineDispatcherFactory
 import io.realm.kotlin.migration.AutomaticSchemaMigration
 import io.realm.kotlin.migration.RealmMigration
 import io.realm.kotlin.types.BaseRealmObject
-import kotlinx.coroutines.CoroutineDispatcher
 import kotlin.reflect.KClass
 
 // TODO Public due to being accessed from `library-sync`
@@ -53,8 +53,8 @@ public open class ConfigurationImpl constructor(
     schema: Set<KClass<out BaseRealmObject>>,
     logConfig: LogConfiguration,
     maxNumberOfActiveVersions: Long,
-    notificationDispatcher: CoroutineDispatcher,
-    writeDispatcher: CoroutineDispatcher,
+    notificationDispatcher: CoroutineDispatcherFactory,
+    writeDispatcher: CoroutineDispatcherFactory,
     schemaVersion: Long,
     schemaMode: SchemaMode,
     private val userEncryptionKey: ByteArray?,
@@ -85,14 +85,13 @@ public open class ConfigurationImpl constructor(
 
     override val mediator: Mediator
 
-    override val notificationDispatcher: CoroutineDispatcher
+    override val notificationDispatcherFactory: CoroutineDispatcherFactory
 
-    override val writeDispatcher: CoroutineDispatcher
+    override val writeDispatcherFactory: CoroutineDispatcherFactory
 
     override val compactOnLaunchCallback: CompactOnLaunchCallback?
 
     override val initialDataCallback: InitialDataCallback?
-
     override fun createNativeConfiguration(): RealmConfigurationPointer {
         val nativeConfig: RealmConfigurationPointer = RealmInterop.realm_config_new()
         return configInitializer(nativeConfig)
@@ -118,8 +117,8 @@ public open class ConfigurationImpl constructor(
         this.mapOfKClassWithCompanion = schema.associateWith { realmObjectCompanionOrThrow(it) }
         this.log = logConfig
         this.maxNumberOfActiveVersions = maxNumberOfActiveVersions
-        this.notificationDispatcher = notificationDispatcher
-        this.writeDispatcher = writeDispatcher
+        this.notificationDispatcherFactory = notificationDispatcher
+        this.writeDispatcherFactory = writeDispatcher
         this.schemaVersion = schemaVersion
         this.schemaMode = schemaMode
         this.compactOnLaunchCallback = compactOnLaunchCallback