+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/RealmImpl.kt
@@ -27,6 +27,7 @@ import io.realm.kotlin.internal.interop.RealmInterop
 import io.realm.kotlin.internal.platform.fileExists
 import io.realm.kotlin.internal.platform.runBlocking
 import io.realm.kotlin.internal.schema.RealmSchemaImpl
+import io.realm.kotlin.internal.util.ManagedCoroutineDispatcher
 import io.realm.kotlin.notifications.RealmChange
 import io.realm.kotlin.notifications.internal.InitialRealmImpl
 import io.realm.kotlin.notifications.internal.UpdatedRealmImpl
@@ -63,14 +64,17 @@ public class RealmImpl private constructor(
 
     private val realmPointerMutex = Mutex()
 
+    public val notificationDispatcher: ManagedCoroutineDispatcher = configuration.notificationDispatcherFactory.create()
+    public val writeDispatcher: ManagedCoroutineDispatcher = configuration.notificationDispatcherFactory.create()
+
     internal val realmScope =
-        CoroutineScope(SupervisorJob() + configuration.notificationDispatcher)
+        CoroutineScope(SupervisorJob() + notificationDispatcher.get())
     private val realmFlow =
         MutableSharedFlow<RealmChange<Realm>>() // Realm notifications emit their initial state when subscribed to
     private val notifier =
-        SuspendableNotifier(this, configuration.notificationDispatcher)
+        SuspendableNotifier(this, notificationDispatcher.get())
     internal val writer =
-        SuspendableWriter(this, configuration.writeDispatcher)
+        SuspendableWriter(this, writeDispatcher.get())
 
     private var _realmReference: AtomicRef<RealmReference> = atomic(LiveRealmReference(this, dbPointer))
     /**
@@ -260,7 +264,8 @@ public class RealmImpl private constructor(
                 realmFlow.emit(UpdatedRealmImpl(this@RealmImpl))
             }
         }
-        // TODO There is currently nothing that tears down the dispatcher
+        notificationDispatcher.closeIfInternal()
+        writeDispatcher.closeIfInternal()
     }
 
     // FIXME Internal method to work around that callback subscription is not freed on GC