+++ packages/library-base/src/darwin/kotlin/io/realm/kotlin/internal/platform/CoroutineUtils.kt
@@ -1,6 +1,6 @@
 package io.realm.kotlin.internal.platform
 
-import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.CloseableCoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.newSingleThreadContext
 import kotlin.coroutines.CoroutineContext
@@ -16,11 +16,11 @@ public actual fun <T> runBlocking(
 /**
  * The default dispatcher for Darwin platforms spawns a new thread with a run loop.
  */
-public actual fun singleThreadDispatcher(id: String): CoroutineDispatcher {
+public actual fun singleThreadDispatcher(id: String): CloseableCoroutineDispatcher {
     return newSingleThreadContext(id)
 }
 
-public actual fun multiThreadDispatcher(size: Int): CoroutineDispatcher {
+public actual fun multiThreadDispatcher(size: Int): CloseableCoroutineDispatcher {
     // TODO https://github.com/realm/realm-kotlin/issues/501
     return singleThreadDispatcher("singleThreadDispatcher")
 }