+++ packages/library-sync/src/commonMain/kotlin/io/realm/kotlin/mongodb/internal/KtorNetworkTransport.kt
@@ -34,11 +34,10 @@ import io.ktor.utils.io.errors.IOException
 import io.realm.kotlin.internal.interop.sync.NetworkTransport
 import io.realm.kotlin.internal.interop.sync.Response
 import io.realm.kotlin.internal.interop.sync.ResponseCallback
+import io.realm.kotlin.internal.util.CoroutineDispatcherFactory
+import io.realm.kotlin.internal.util.ManagedCoroutineDispatcher
 import io.realm.kotlin.mongodb.AppConfiguration.Companion.DEFAULT_AUTHORIZATION_HEADER_NAME
-import kotlinx.coroutines.CancellationException
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.async
+import kotlinx.coroutines.*
 import kotlin.collections.set
 
 public class KtorNetworkTransport(
@@ -47,10 +46,11 @@ public class KtorNetworkTransport(
     // FIXME Rework timeout to take a Duration instead
     //  https://github.com/realm/realm-kotlin/issues/408
     timeoutMs: Long,
-    private val dispatcher: CoroutineDispatcher,
+    dispatcher: CoroutineDispatcherFactory,
     logger: Logger? = null,
 ) : NetworkTransport {
 
+    private val dispatcher: ManagedCoroutineDispatcher = dispatcher.create()
     private val clientCache: HttpClientCache = HttpClientCache(timeoutMs, logger)
 
     @Suppress("ComplexMethod", "TooGenericExceptionCaught")
@@ -62,7 +62,7 @@ public class KtorNetworkTransport(
         callback: ResponseCallback,
     ) {
         val client = clientCache.getClient()
-        CoroutineScope(dispatcher).async {
+        CoroutineScope(dispatcher.get()).async {
             val response = try {
                 val requestBuilderBlock: HttpRequestBuilder.() -> Unit = {
                     headers {
@@ -128,6 +128,7 @@ public class KtorNetworkTransport(
 
     override fun close() {
         clientCache.close()
+        dispatcher.closeIfInternal()
     }
 
     private suspend fun processHttpResponse(response: HttpResponse): Response {